Clarify a previously implicit goal of the workshop
which is to set up the learner with a proper local dev env/workflow for
future success.

A discussion in issue #7 regarding possible interactive/online format of
the workshop material led to this clarification.

There are other ancillary reasons why it is the way it is, discussed in
the issue, and x-posted below for the permanent record:

cf. https://github.com/adityaathalye/clojure-by-example/issues/7

 dijonkitchen commented 27 days ago

Maybe consider using https://github.com/viebel/klipse for faster
feedback right on the same page instead of repl.it or static,
pre-formatted text?

Might help let people play around with the code and see how it works.

 adityaathalye commented yesterday

Thank you for the suggestion.

There are a few considerations for the choice of development
environment:

  1. This workshop targets practicing programmers, and not complete
beginners. Our opinion is that getting started with standard developer
tooling is part of the learning. We don't want to recommend Emacs (most
of us use Emacs), because tha't too alien for most people. So far,
Intellij + Cursive has been a rock-solid, if heavy to set up, option.

  2. High quality developer tooling is crucial for a good learning
experience --- static checks, linting, paredit/parinfer support,
super-fast feedback etc. help immensely. The standard developer tooling
sees heavy investment in imporving these creature comforts. The workshop
participants also expect tools of the quality they are accustomed to in
their normal day to day work.

  3. Avoid Internet connectivity when teaching in-person --- this is a
failure mode / source of degraded experience we wish to avoid when time
is of the essence. We are able to avoid it, with a combination of
project-local .m2, which we can share via USB stick, and previously
configured local dev setup.

  4. Set up for future success. We hope programmers would want to keep
going after the workshop, and start hobby projects. Having a local dev
environment already set up removes the friction of having to switch
tools / interaction models.

  5. Maintainig support for multiple dev environments is overhead we
can't take. Currently, we test the flow of the workshop on (Mac, Linux,
Windows) x (1 IDE) x (1 local JVM). We don't have the bandwidth to
ensure the learning experience is reliable on more platforms. If we
choose an online dev tool, we also have to worry about browser
compatibility issues and the necessarily sandboxed / opaque runtime of
the online service.