Fix behaviour of "standard input" on Unix-like operating systems
Previously the ClojureCLR `*in*` (standard input) stream wouldn't behave
correctly on Unix-like operating systems such as macOS and Linux.  When
used to accept user input, it would incorrectly handle backspacing and
the use of the cursor/arrow keys.

As ClojureCLR's REPL reads from `*in*` it too was affected by this bug.

This bug also prevented the REPL from being able to work with other
command line tools such as `rlwrap`.

Switching from `Console.In` to `Console.OpenStandardInput` resolved the
issue.

This commit also updates the other standard streams (standard error and
standard output) so that they all work the same way.